Refactor overall: Added structure to the project with directories | Updated from Bootstrap v3.3.6 to v4.5.2 | Added header and footer files to allow for single page updates on content found on multiple pages | Preliminary UI includes admin specified puzzles on index.php, category.php has been added, puzzle.php (formerly getInfo.php) and createPuzzle.php (formerly index.html) have been completely overhauled | NOTE: The solution puzzle (bottom of puzzle.php) does not currently render; this will be amended in future iterations.
