+++ blank_page.php
@@ -1,0 +1,12 @@
+<?php 
+
+    $page_title = '';
+    include('inc/header.php');
+
+?>
+
+<div class="container">
+
+</div>
+
+<?php include('inc/footer.php'); ?>
