+++ category.php
@@ -1,0 +1,51 @@
+<?php 
+
+    $page_title = '';
+    include('inc/header.php');
+
+?>
+
+<div class="container">
+
+    <div class="card mt-4">
+        <div class="card-header">NBA Teams</div>
+    </div>
+
+    <div class="card mt-4">
+        <h5 class="card-header"><a href="puzzle.php" class="card-link">Timberwolves</a></h5>
+        <div class="card-body">
+            <p class="card-text">MN based team whose only good seasons were with Garnett!</p>
+        </div>
+    </div>
+
+    <div class="card mt-4">
+        <h5 class="card-header"><a href="puzzle.php" class="card-link">Bulls</a></h5>
+        <div class="card-body">
+            <p class="card-text">This NBA team is based out of Chicago, IL and they stole Zack Lavine from the Timberwolves!</p>
+        </div>
+    </div>
+
+    <div class="card mt-4">
+        <h5 class="card-header"><a href="puzzle.php" class="card-link">Warriors</a></h5>
+        <div class="card-body">
+            <p class="card-text">This NBA team has the legenday Steph Curry on their roster.</p>
+        </div>
+    </div>
+
+    <div class="card mt-4">
+        <h5 class="card-header"><a href="puzzle.php" class="card-link">Bucks</a></h5>
+        <div class="card-body">
+            <p class="card-text">This NBA team is better than the Timberwolves but they are still overrated!!</p>
+        </div>
+    </div>
+
+    <div class="card mt-4">
+        <h5 class="card-header"><a href="puzzle.php" class="card-link">Heat</a></h5>
+        <div class="card-body">
+            <p class="card-text">Jimmy Buckets!!!</p>
+        </div>
+    </div>
+    
+</div>
+
+<?php include('inc/footer.php'); ?>
