+++ inc/footer.php
@@ -1,0 +1,14 @@
+    </main>
+
+    <footer class="bg-dark">
+       <div class="container text-white">
+            silc | Word Find <?php echo date('Y');  ?>
+       </div>
+    </footer>
+
+</body>
+
+<script src="js/main.js"></script>
+
+</body>
+</html>
