+++ index.php
@@ -1,0 +1,181 @@
+<?php 
+
+    $page_title = 'Word Find';
+    include('inc/header.php');
+
+?>
+
+<div class="container">
+
+    <div class="card mt-4">
+        <div class="card-header">Popular Categories</div>
+    </div>
+
+    <!-- Row 1 -->
+    <div class="row mt-4">
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">NBA Teams</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Timberwolves</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Bulls</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Warriors</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Bucks</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Heat</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">NFL Teams</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Vikings</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Packers</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Bears</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Dolphins</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Cowboys</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">MLB Teams</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Twins</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Cubs</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Blue Jays</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">White Sox</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Giants</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">NHL Teams</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Oilers</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Maple Leafs</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Stars</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Rangers</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Wild</a></p>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <!-- Row 2 -->
+    <div class="row mt-4">
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">Television Shows</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Star Trek</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">The Unit</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">SG1</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Stargate Atlantis</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Batman Beyond</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">Sitcoms</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Friends</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Frasier</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Two and a Half Men</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">The Bing Bang Theory</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">How I met Your Mother</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">Movies</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">John Wick</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">The Matrix</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Star Wars</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Harry Potter</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Universal Soldier</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+            <h5 class="card-header"><a href="category.php" class="card-link">Animals</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Dogs</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Most Common Pets</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Safari Animals</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Farm Animals</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Extinct</a></p>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <!-- Row 3 -->
+    <div class="row mt-4">
+        <div class="col-sm-3">
+            <div class="card">
+            <h5 class="card-header"><a href="category.php" class="card-link">Music</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">NF Songs</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Blink 182</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Famous Bands</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Musical Terms</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">MGK songs</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">Disney</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Mandalorian</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Disney Villians</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Disney Princesses</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Animated Movies</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Duck Tales</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">Science</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Human Bones</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Algebra</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Metals</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Gases</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Igneous Rock</a></p>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-3">
+            <div class="card">
+                <h5 class="card-header"><a href="category.php" class="card-link">Food and Drink</a></h5>
+                <div class="card-body">
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Pizza Toppings</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Fast Food</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Fruits</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Vegetables</a></p>
+                    <p class="card-text"><a href="puzzle.php" class="card-link">Herbs & Spices</a></p>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+
+<?php include('inc/footer.php'); ?>
