--- style.css
@@ -1,102 +1,0 @@
-﻿html {
-	font-family: Helvetica, Arial, sans-serif;
-}
-.jumbotron {
-	background-image: url("silcHeader.png");
-	-webkit-background-size: 100% 100%;
-	-moz-background-size: 100% 100%;
-	-o-background-size: 100% 100%;
-	background-size: 100% 100%;
-	height: 179px;
-}
-
-table, th, tr, td {
-
-	border-collapse: collapse;
-	padding: 5px;
-	font-size: 36px;
-
-}
-.totals {
-	border-style: none;
-	border-bottom: none;
-	border-top: none;
-	font-size: 1.2em;
-}
-.pass {
-	color: green;
-}
-
-.fail {
-	color: red;
-}
-body { 
-	font-family: "Helvetica", Verdana, sans-serif;
-	font-size: 1.1em;
-	background-color: white;
-	margin-left: 15%;
-	margin-right: 15%;
-	margin-top: 0%;
-	border-left: 1px solid black;
-	border-right: 1px solid black;
-	border-bottom: 1px solid black;
-	border-radius: 5px;
-
-}
-h1 {
-	font-family: "Helvetica", Verdana, sans-serif; 
-	font-size: 2em;
-	font-weight: bold;
-	color: black;
-	text-align: center;
-	background-color:#337ab7;
-}
-.tabs {
-	margin: 0;
-	padding: 0;
-}
-.tabs li {
-	float: left;
-	list-style: none;
-	padding: 0;
-	margin: 0;
-}
-.tabs a {
-	display: block;
-	text-decoration: none;
-	padding: 3px 5px;
-	margin-right: 10px;
-	margin-bottom: -1px;
-	border: 1px solid black;
-	background-color: grey;
-	border-top-right-radius: 0.5em;
-	border-top-left-radius: 0.5em;
-}
-.tabs .fore {
-	background-color: blue;
-	position: relative;
-	border-bottom: 1px solid blue;
-	color: white;
-}
-.panels {
-	clear: both;
-	width: 50%;
-	border: 1px solid blue;
-	margin-bottom: 25px;
-	padding: 10px;
-}
-.panel p {
-	color: black;
-}
-.rectangle{
-	visibility: visible;
-	height: 42px;
-	background: none;
-	border-radius: 20px;
-	border: 2px solid red;
-	-webkit-transform-origin: 21px 21px ;
-	-ms-transform-origin: 21px 21px;
-	-o-transform-origin: 21px 21px;
-	transform-origin: 21px 21px;
-
-}
